+++ src/view/index.ts
@@ -4,3 +4,4 @@ export * from './dom-renderer';
 export * from './selection-handler';
 export * from './web-component';
 export * from './media';
+export * from './menu-position';